pyproject.toml
[tool.poetry]
name = "fastapi_template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
black = "^21.11b1"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.910"
tox = "^3.24.4"
oauth2client = "^4.1.3"
firebase-admin = "^5.3.0"
tzdata = "^2022.2"
google-cloud-storage = "^2.5.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
